add commandline
